refactor: Factor out most ops from executor
Handling control flow must be solved first before last op gets factored out.
