+++ src/machine/ops/index.js
@@ -1,10 +1,17 @@
 module.exports = {
-  LdaSmi: require('./LdaSmi'),
+  CallProperty1: require('./CallProperty1'),
+  CallUndefinedReceiver: require('./CallUndefinedReceiver'),
+  Debugger: require('./Debugger'),
+  Jump: require('./Jump'),
   LdaConstant: require('./LdaConstant'),
+  LdaGlobal: require('./LdaGlobal'),
+  LdaNamedProperty: require('./LdaNamedProperty'),
+  Ldar: require('./Ldar'),
+  LdaSmi: require('./LdaSmi'),
   LdaUndefined: require('./LdaUndefined'),
   LdaZero: require('./LdaZero'),
-  Return: require('./Return'),
   MulSmi: require('./MulSmi'),
+  Return: require('./Return'),
   Star: require('./Star'),
   TestLessThan: require('./TestLessThan'),
 };
